pyproject.toml
[tool.poetry]
name = "trees"
version = "0.1.0"
description = ""
authors = ["vova <vladimir1012kov@gmail.com>"]
packages = [{include = "main_code"}]

[tool.poetry.scripts]
first = "main_code.scripts.test_work:main"

[tool.poetry.dependencies]
python = "^3.8"
hexlet-immutable-fs-trees = "^0.1.7"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
